Added Dakota's version of 9.10
I took a different approach that works. The one thing that didn't work in zyLabs was the "Tests that the default constructor initializes attributes correctly". I don't know what that's referring to, but the program functions correctly outside of zyLabs.

Screenshot of zyLabs error: http://logicalfunctionality.com/CIS1415/Capture.PNG